requirements.txt
google-genai>=0.1.0
python-dotenv>=1.0.0
ipython>=8.0.0
tomli>=2.0.0
tomli-w>=1.0.0
rich>=10.0.0
Flask>=2.3.0
icecream>=2.0.0
openai>=1.0.0
opencv-python>=4.8.0
pillow>=10.0.0
pypdf>=6.0.0
